chore: not use js entries, but use ts entries with require
