--- pkg/api/bin/lambda.js
@@ -1,2 +1,0 @@
-require('source-map-support').install()
-module.exports = require('../build/lambda.js')
