+++ pkg/api/package.json
@@ -5,18 +5,18 @@
   "scripts": {
     "migrate:dev": "pnpm run _:build && pnpm run _:migrate:dev",
     "migrate:reset": "pnpm run _:build && pnpm run _:migrate:reset",
-    "start": "cross-env NODE_ENV=production node --require=source-map-support/register build/index.js",
+    "start": "cross-env NODE_ENV=production node build/index.js",
     "prisma:prod": "pnpm run _:build && pnpm run _:prisma:prod",
     "prisma:seed": "pnpm run _:build && pnpm run _:prisma:seed",
     "prisma:seed:dev": "pnpm run _:build && pnpm run prisma:seed:dev",
     "_:migrate:dev": "prisma migrate dev",
     "_:migrate:reset": "prisma migrate reset",
-    "_:prisma:prod": "node --require=source-map-support/register ./prisma/build/index.js --prod",
-    "_:prisma:seed": "node --require=source-map-support/register ./prisma/build/index.js --seed",
+    "_:prisma:prod": "node ./prisma/build/index.js --prod",
+    "_:prisma:seed": "node ./prisma/build/index.js --seed",
     "_:prisma:seed:dev": "pnpm run _:prisma:seed -- dev",
     "_:dev": "run-p -l _:dev:*",
     "_:dev:build": "pnpm run migrate:dev && pnpm run _:prisma:seed:dev && run-p -l \"_:build:* -- --watch\"",
-    "_:dev:serve": "node-dev src/entries/index.ts",
+    "_:dev:serve": "node-dev --respawn build/index.js",
     "_:dev:aspida": "aspida --watch",
     "_:dev:frourio": "onchange -i -k -d 1000 './api/**/*' -e '**/$*.ts' -- frourio",
     "_:build:index": "ts-node -T ./node_modules/@violet/scripts/build-files.ts --from-dir=./src/entries --to-dir=./build --target=node14",
