--- pkg/lambda/conv2img/bin/index.js
@@ -1,2 +1,0 @@
-require('source-map-support').install()
-module.exports = require('../build/index.js')
